Change definition for v-model of fields using reactive store
