+++ src/components/ADempiere/ContextMenu/contextMenuMixin.js
@@ -174,7 +174,7 @@ export const contextMixin = {
     },
     getOldRouteOfWindow() {
       if (this.panelType === 'window') {
-        const oldRoute = this.$store.state.windowControl.windowOldRoute
+        const oldRoute = this.$store.state.window.windowOldRoute
         if (!this.isEmptyValue(oldRoute.query.action) && oldRoute.query.action !== 'create-new' && this.$route.query.action === 'create-new') {
           return oldRoute
         }
