+++ src/components/ADempiere/Field/FieldMixin.js
@@ -13,17 +13,35 @@ export const fieldMixin = {
   },
   data() {
     // value render
-    let value = this.metadata.value
+    let value1 = this.metadata.value
     if (this.metadata.inTable) {
-      value = this.valueModel
+      value1 = this.valueModel
     }
     return {
-      value
+      value1
     }
   },
   computed: {
     isDisabled() {
       return Boolean(this.metadata.readonly || this.metadata.disabled)
+    },
+    value: {
+      get() {
+        // console.log(this.$store)
+        return this.$store.getters.getValueOfField({
+          containerUuid: this.metadata.containerUuid,
+          columnName: this.metadata.columnName
+        })
+      },
+      set(value) {
+        // console.log(value)
+        this.$store.commit('updateValueOfField', {
+          parentUuid: this.metadata.parentUuid,
+          containerUuid: this.metadata.containerUuid,
+          columnName: this.metadata.columnName,
+          value: value
+        })
+      }
     }
   },
   async created() {
@@ -41,7 +59,22 @@ export const fieldMixin = {
       this.requestFocus()
     }
   },
+  watch: {
+    valueModel(value) {
+      if (this.metadata.inTable) {
+        this.value = this.parseValue(value)
+      }
+    },
+    'metadata.value'(value) {
+      if (!this.metadata.inTable) {
+        this.value = this.parseValue(value)
+      }
+    }
+  },
   methods: {
+    parseValue(value) {
+      return value
+    },
     /**
      * Set focus if handle focus attribute is true
      */
@@ -120,42 +153,12 @@ export const fieldMixin = {
      * @param {string} label, or displayColumn to show in select
      */
     handleChange(value, valueTo = undefined, label = undefined) {
-      let newValue = value
-      let isSendCallout = true
-      let isSendToServer = true
-      let isChangedOldValue = false
-      if (value === 'NotSend') {
-        newValue = this.value
-        if (this.componentPath === 'FieldYesNo') {
-          isChangedOldValue = true
-          newValue = Boolean(newValue)
-        }
-        isSendToServer = false
-        isSendCallout = false
-      }
-      if (this.metadata.isAdvancedQuery) {
-        isSendCallout = false
-      }
-
-      const sendParameters = {
-        parentUuid: this.metadata.parentUuid,
-        containerUuid: this.metadata.containerUuid,
-        field: this.metadata,
-        panelType: this.metadata.panelType,
-        columnName: this.metadata.columnName,
-        newValue,
-        valueTo,
-        isAdvancedQuery: this.metadata.isAdvancedQuery,
-        isSendToServer,
-        isSendCallout,
-        isChangedOldValue
-      }
       // Global Action performed
       if (this.metadata.handleActionPerformed) {
         this.$store.dispatch('notifyActionPerformed', {
           containerUuid: this.metadata.containerUuid,
           columnName: this.metadata.columnName,
-          value: newValue
+          value: value
         })
       }
 
@@ -163,20 +166,13 @@ export const fieldMixin = {
       if (this.metadata.isCustomField) {
         return
       }
-
-      if (this.metadata.inTable) {
-        this.$store.dispatch('notifyCellTableChange', {
-          ...sendParameters,
-          keyColumn: this.metadata.keyColumn,
-          tableIndex: this.metadata.tableIndex,
-          rowKey: this.metadata.rowKey
-        })
-      } else {
-        this.$store.dispatch('notifyFieldChange', {
-          ...sendParameters,
-          displayColumn: label
-        })
-      }
+      this.$store.dispatch('notifyFieldChange', {
+        parentUuid: this.metadata.parentUuid,
+        containerUuid: this.metadata.containerUuid,
+        panelType: this.metadata.panelType,
+        columnName: this.metadata.columnName,
+        field: this.metadata
+      })
     }
   }
 }
