+++ src/store/modules/ADempiere/browser.js
@@ -1,188 +1,181 @@
-import { getBrowser as getBrowserMetadata } from '@/api/ADempiere/dictionary'
-import { showMessage } from '@/utils/ADempiere/notification'
+import { getBrowserSearch } from '@/api/ADempiere/browser'
 import { isEmptyValue } from '@/utils/ADempiere/valueUtils'
-import { generateField } from '@/utils/ADempiere/dictionaryUtils'
-import router from '@/router'
+import { parseContext } from '@/utils/ADempiere/contextUtils'
+import { showMessage } from '@/utils/ADempiere/notification'
+import { fieldIsDisplayed } from '@/utils/ADempiere/dictionaryUtils'
 import language from '@/lang'
 
-const browser = {
-  state: {
-    browser: []
-  },
-  mutations: {
-    addBrowser(state, payload) {
-      state.browser.push(payload)
-    },
-    dictionaryResetCacheBrowser(state) {
-      state.browser = []
-    },
-    changeBrowserAttribute(state, payload) {
-      let value = payload.attributeValue
-      if (payload.attributeNameControl) {
-        value = payload.browser[payload.attributeNameControl]
-      }
-      payload.browser[payload.attributeName] = value
-    }
-  },
+const browserControl = {
   actions: {
-    getBrowserFromServer({ commit, dispatch }, {
+    browserActionPerformed({ dispatch, getters }, {
+      parentUuid,
       containerUuid,
-      routeToDelete
+      tableName,
+      columnName,
+      field,
+      value
     }) {
-      return new Promise(resolve => {
-        getBrowserMetadata(containerUuid)
-          .then(browserResponse => {
-            const panelType = 'browser'
-            const additionalAttributes = {
-              containerUuid,
-              panelType,
-              isEvaluateValueChanges: true
-            }
-            const {
-              query,
-              whereClause,
-              process
-            } = browserResponse
-
-            //  Convert from gRPC
-            const fieldsRangeList = []
-            let isShowedCriteria = false
-            let awaitForValues = 0
-            let fieldsList = browserResponse.fieldsList.map((fieldItem, index) => {
-              const someAttributes = {
-                ...additionalAttributes,
-                fieldListIndex: index
-              }
-              const field = generateField({
-                fieldToGenerate: fieldItem,
-                moreAttributes: someAttributes,
-                isSOTrxMenu: routeToDelete.meta.isSOTrx
-              })
-              // Add new field if is range number
-              if (field.isRange && field.componentPath === 'FieldNumber') {
-                const fieldRange = generateField({
-                  fieldToGenerate: fieldItem,
-                  moreAttributes: someAttributes,
-                  typeRange: true
-                })
-                if (!isEmptyValue(fieldRange.value)) {
-                  fieldRange.isShowedFromUser = true
-                }
-                fieldsRangeList.push(fieldRange)
-              }
-
-              // Only isQueryCriteria fields with values, displayed in main panel
-              if (field.isQueryCriteria) {
-                if (field.isSQLValue) {
-                  isShowedCriteria = true
-                  field.isShowedFromUser = true
-                  awaitForValues++
-                }
-                if (query.includes(`@${field.columnName}@`) ||
-                  query.includes(`@${field.columnName}_To@`) ||
-                  whereClause.includes(`@${field.columnName}@`) ||
-                  whereClause.includes(`@${field.columnName}_To@`)) {
-                  field.isMandatory = true
-                  field.isMandatoryFromLogic = true
-                  field.isShowedFromUser = true
-                }
-
-                if (isEmptyValue(field.value)) {
-                  // isMandatory params to showed search criteria
-                  if (field.isMandatory || field.isMandatoryFromLogic) {
-                    isShowedCriteria = true
-                  }
-                } else {
-                  // with value
-                  field.isShowedFromUser = true
-                }
-              }
-
-              return field
-            })
-            fieldsList = fieldsList.concat(fieldsRangeList)
-
-            // Panel for save on store
-            const newBrowser = {
-              ...browserResponse,
-              containerUuid,
-              fieldList: fieldsList,
-              panelType,
-              // app attributes
-              awaitForValues, // control to values
-              awaitForValuesToQuery: awaitForValues, // get values from request search
-              isShowedCriteria,
-              isShowedTotals: true
-            }
-
-            commit('addBrowser', newBrowser)
-            dispatch('addPanel', newBrowser)
-
-            resolve(newBrowser)
-
-            // Convert from gRPC process list
-            const actions = []
-            if (process) {
-              actions.push({
-                type: 'process',
-                panelType: 'process',
-                uuid: process.uuid,
-                name: process.name,
-                description: process.description,
-                isReport: process.isReport,
-                isDirectPrint: process.isDirectPrint,
-                containerUuidAssociated: containerUuid,
-                panelTypeAssociated: panelType
-              })
-              // TODO: No list of parameters
-              // // add process associated in vuex store
-              // dispatch('addProcessAssociated', {
-              //   processToGenerate: process,
-              //   containerUuidAssociated: containerUuid
-              // })
-            }
-            // Add process menu
-            dispatch('setContextMenu', {
-              containerUuid,
-              actions
-            })
-          })
-          .catch(error => {
-            router.push({ path: '/dashboard' })
-            dispatch('tagsView/delView', routeToDelete)
-            showMessage({
-              message: language.t('login.unexpectedError'),
-              type: 'error'
-            })
-            console.warn(`Dictionary Browser - Error ${error.code}: ${error.message}.`)
-          })
+      console.log({
+        parentUuid,
+        containerUuid,
+        tableName,
+        columnName,
+        field,
+        value
       })
-    },
-    changeBrowserAttribute({ commit, getters }, {
-      containerUuid,
-      browser,
-      attributeName,
-      attributeNameControl,
-      attributeValue
-    }) {
-      if (isEmptyValue(browser)) {
-        browser = getters.getBrowser(containerUuid)
+      const fieldsEmpty = getters.getFieldListEmptyMandatory({
+        containerUuid,
+        fieldsList: getters.getFieldsListFromPanel(containerUuid)
+      })
+      if (isEmptyValue(fieldsEmpty)) {
+        showMessage({
+          message: language.t('notifications.mandatoryFieldMissing') + fieldsEmpty,
+          type: 'info'
+        })
+        return
       }
-      commit('changeBrowserAttribute', {
-        browser,
-        attributeName,
-        attributeValue,
-        attributeNameControl
+      // Validate if a field is called and visible
+      if (fieldIsDisplayed(field)) {
+        let isReadyForQuery = true
+        if (field.isSQLValue) {
+          const panel = getters.getPanel(containerUuid)
+          let awaitForValuesToQuery = panel.awaitForValuesToQuery
+          awaitForValuesToQuery--
+          dispatch('changeBrowserAttribute', {
+            containerUuid,
+            attributeName: 'awaitForValuesToQuery',
+            attributeValue: awaitForValuesToQuery
+          })
+          if (awaitForValuesToQuery === 0) {
+            if (panel.isShowedCriteria) {
+              dispatch('changeBrowserAttribute', {
+                containerUuid,
+                attributeName: 'isShowedCriteria',
+                attributeValue: false
+              })
+            }
+          } else if (awaitForValuesToQuery > 0) {
+            isReadyForQuery = false
+          }
+        }
+        if (isReadyForQuery && !field.dependentFieldsList.length) {
+          dispatch('getBrowserSearch', {
+            containerUuid,
+            isClearSelection: true
+          })
+        }
+      }
+    },
+    // Search with query criteria
+    getBrowserSearch({ dispatch, rootGetters }, {
+      containerUuid,
+      isClearSelection = false
+    }) {
+      showMessage({
+        title: language.t('notifications.loading'),
+        message: language.t('notifications.searching'),
+        type: 'info'
       })
-    }
-  },
-  getters: {
-    getBrowser: (state) => (browserUuid) => {
-      return state.browser.find(
-        item => item.uuid === browserUuid
-      )
+      const allData = rootGetters.getDataRecordAndSelection(containerUuid)
+      // deletes the data from the container to replace it and to report the searches in the table
+      dispatch('deleteRecordContainer', {
+        viewUuid: containerUuid
+      })
+
+      const browser = rootGetters.getBrowser(containerUuid)
+      // parameters isQueryCriteria
+      const parametersList = rootGetters.getParametersToServer({
+        containerUuid,
+        fieldList: browser.fieldList
+      })
+
+      let parsedQuery = browser.query
+      if (!isEmptyValue(parsedQuery) && parsedQuery.includes('@')) {
+        parsedQuery = parseContext({
+          containerUuid,
+          value: parsedQuery,
+          isBooleanToString: true
+        }).value
+      }
+
+      let parsedWhereClause = browser.whereClause
+      if (!isEmptyValue(parsedWhereClause) && parsedWhereClause.includes('@')) {
+        parsedWhereClause = parseContext({
+          containerUuid,
+          value: parsedWhereClause,
+          isBooleanToString: true
+        }).value
+      }
+
+      let nextPageToken
+      if (!isEmptyValue(allData.nextPageToken)) {
+        nextPageToken = allData.nextPageToken + '-' + allData.pageNumber
+      }
+
+      // Add validation compare browserSearchQueryParameters
+      return getBrowserSearch({
+        uuid: containerUuid,
+        query: parsedQuery,
+        whereClause: parsedWhereClause,
+        orderByClause: browser.orderByClause,
+        parametersList,
+        nextPageToken: nextPageToken
+      })
+        .then(browserSearchResponse => {
+          const recordsList = browserSearchResponse.recordsList.map(itemRecord => {
+            return {
+              ...itemRecord.values,
+              // datatables attributes
+              isNew: false,
+              isEdit: false,
+              isSelected: false,
+              isReadOnlyFromRow: false
+            }
+          })
+
+          let selection = allData.selection
+          if (isClearSelection) {
+            selection = []
+          }
+
+          let token = browserSearchResponse.nextPageToken
+          if (token !== undefined) {
+            token = token.slice(0, -2)
+          }
+
+          dispatch('setRecordSelection', {
+            containerUuid,
+            record: recordsList,
+            pageNumber: rootGetters.getPageNumber(containerUuid),
+            selection: selection,
+            recordCount: browserSearchResponse.recordCount,
+            nextPageToken: token
+          })
+          showMessage({
+            title: language.t('notifications.succesful'),
+            message: language.t('notifications.succcessSearch'),
+            type: 'success'
+          })
+          return recordsList
+        })
+        .catch(error => {
+          // Set default registry values so that the table does not say loading,
+          // there was already a response from the server
+          dispatch('setRecordSelection', {
+            containerUuid,
+            panelType: 'browser'
+          })
+
+          showMessage({
+            title: language.t('notifications.error'),
+            message: language.t('notifications.errorSearch'),
+            summary: error.message,
+            type: 'error'
+          })
+          console.warn(`Error getting browser search: ${error.message}. Code: ${error.code}.`)
+        })
     }
   }
 }
 
-export default browser
+export default browserControl
