+++ src/store/modules/ADempiere/window.js
@@ -1,437 +1,1113 @@
-import {
-  getWindow as getWindowMetadata,
-  getTab as getTabMetadata
-} from '@/api/ADempiere/dictionary'
+import { createEntity, updateEntity, deleteEntity, rollbackEntity } from '@/api/ADempiere/persistence'
+import { getReferencesList } from '@/api/ADempiere/values'
+import { convertObjectToArrayPairs, isEmptyValue } from '@/utils/ADempiere/valueUtils'
+import { fieldIsDisplayed } from '@/utils/ADempiere/dictionaryUtils'
+import { parseContext } from '@/utils/ADempiere/contextUtils'
 import { showMessage } from '@/utils/ADempiere/notification'
-import { isEmptyValue } from '@/utils/ADempiere/valueUtils'
 import language from '@/lang'
 import router from '@/router'
-import { generateField } from '@/utils/ADempiere/dictionaryUtils'
-import { getFieldTemplate } from '@/utils/ADempiere/lookupFactory'
 
-const initStateWindow = {
-  window: [],
-  windowIndex: 0
+const initStateWindowControl = {
+  inCreate: [],
+  references: [],
+  windowOldRoute: {
+    path: '',
+    fullPath: '',
+    query: {}
+  },
+  dataLog: {}, // { containerUuid, recordId, tableName, eventType }
+  tabSequenceRecord: [],
+  totalResponse: 0,
+  totalRequest: 0
 }
 
-const window = {
-  state: initStateWindow,
+const windowControl = {
+  state: initStateWindowControl,
   mutations: {
-    addWindow(state, payload) {
-      state.window.push(payload)
-      state.windowIndex++
+    addInCreate(state, payload) {
+      state.inCreate.push(payload)
     },
-    dictionaryResetCacheWindow(state) {
-      state = initStateWindow
+    deleteInCreate(state, payload) {
+      state.inCreate = state.inCreate.filter(item => item.containerUuid !== payload.containerUuid)
     },
-    setCurrentTab(state, payload) {
-      payload.window.currentTab = payload.tab
-      payload.window.currentTabUuid = payload.tab.uuid
+    addReferencesList(state, payload) {
+      state.references.push(payload)
     },
-    changeWindowAttribute(state, payload) {
-      let value = payload.attributeValue
-      if (payload.attributeNameControl) {
-        value = payload.window[payload.attributeNameControl]
-      }
-      payload.window[payload.attributeName] = value
+    setDataLog(state, payload) {
+      state.dataLog = payload
     },
-    changeTabAttribute(state, payload) {
-      let value = payload.attributeValue
-      if (payload.attributeNameControl) {
-        value = payload.tab[payload.attributeNameControl]
-      }
-      payload.tab[payload.attributeName] = value
+    setWindowOldRoute(state, payload) {
+      state.windowOldRoute = payload
+    },
+    setTabSequenceRecord(state, payload) {
+      state.tabSequenceRecord = payload
+    },
+    setTotalResponse(state, payload) {
+      state.totalResponse = payload
+    },
+    setTotalRequest(state, payload) {
+      state.totalRequest = payload
+    },
+    resetStateWindowControl(state) {
+      state = initStateWindowControl
     }
   },
   actions: {
-    getWindowFromServer({ commit, state, dispatch }, {
-      windowUuid,
-      routeToDelete
-    }) {
-      return getWindowMetadata(windowUuid)
-        .then(responseWindow => {
-          const firstTabTableName = responseWindow.tabsList[0].tableName
-          const firstTabUuid = responseWindow.tabsList[0].uuid
-          const tabsListParent = []
-          const tabsListChildren = []
-
-          const tabsSequence = []
-          // TODO Add source tab on the server for tabs Translation and Sort
-          const tabs = responseWindow.tabsList.filter(itemTab => {
-            if (itemTab.isSortTab) {
-              // TODO: Add convert tab as process function
-              tabsSequence.push({
-                uuid: itemTab.uuid,
-                id: itemTab.id,
-                parentUuid: windowUuid,
-                containerUuid: itemTab.uuid,
-                parentTabUuid: itemTab.parentTabUuid,
-                panelType: 'window',
-                type: 'sequence',
-                isSortTab: itemTab.isSortTab,
-                name: itemTab.name,
-                description: itemTab.description,
-                tableName: itemTab.tableName,
-                sortOrderColumnName: itemTab.sortOrderColumnName, // order column
-                sortYesNoColumnName: itemTab.sortYesNoColumnName // included column
-              })
-            }
-            // TODO: Add support to isAdvancedTab, isTranslationTab and isHasTree
-            return !itemTab.isTranslationTab
-          }).map((tabItem, index, list) => {
-            // let tab = tabItem
-            const tab = {
-              ...tabItem,
-              containerUuid: tabItem.uuid,
-              parentUuid: windowUuid,
-              windowUuid,
-              tabGroup: tabItem.fieldGroup,
-              firstTabUuid,
-              // relations
-              isParentTab: Boolean(firstTabTableName === tabItem.tableName),
-              // app properties
-              isAssociatedTabSequence: false, // show modal with order tab
-              isShowedRecordNavigation: !(tabItem.isSingleRow),
-              isLoadFieldsList: false,
-              index
-            }
-            delete tab.processesList
-
-            // action is dispatch used in vuex
-            let actions = []
-            actions.push({
-              // action to set default values and enable fields not isUpdateable
-              name: language.t('window.newRecord'),
-              processName: language.t('window.newRecord'),
-              type: 'dataAction',
-              action: 'resetPanelToNew',
-              uuidParent: windowUuid,
-              disabled: !tab.isInsertRecord || tab.isReadOnly
-            }, {
-              // action to delete record selected
-              name: language.t('window.deleteRecord'),
-              processName: language.t('window.deleteRecord'),
-              type: 'dataAction',
-              action: 'deleteEntity',
-              uuidParent: windowUuid,
-              disabled: tab.isReadOnly
-            }, {
-              // action to undo create, update, delete record
-              name: language.t('data.undo'),
-              processName: language.t('data.undo'),
-              type: 'dataAction',
-              action: 'undoModifyData',
-              uuidParent: windowUuid,
-              disabled: false
-            }, {
-              name: language.t('data.lockRecord'),
-              processName: language.t('data.lockRecord'),
-              type: 'dataAction',
-              action: 'lockRecord',
-              disabled: false,
-              hidden: true,
-              tableName: '',
-              recordId: null
-            }, {
-              name: language.t('data.unlockRecord'),
-              processName: language.t('data.unlockRecord'),
-              type: 'dataAction',
-              action: 'unlockRecord',
-              disabled: false,
-              hidden: true,
-              tableName: '',
-              recordId: null
-            })
-
-            if (tab.isSortTab) {
-              const tabParent = list.find(itemTab => itemTab.tableName === tab.tableName && !itemTab.isSortTab)
-              if (tabParent) {
-                tab.tabAssociatedUuid = tabParent.uuid // tab source uuid
-                tab.tabAssociatedName = tabParent.name // tab source name
-              }
-            } else {
-              // add tabs sequence associated as process in tab source
-              let orderTabs = tabsSequence.filter(itemTab => itemTab.tableName === tab.tableName)
-              if (orderTabs.length) {
-                orderTabs = orderTabs.map(itemTab => {
-                  return {
-                    ...itemTab,
-                    // appication attributes
-                    tabAssociatedUuid: tab.uuid, // tab source
-                    tabAssociatedName: tab.name, // tab source
-                    action: 'orderSequence',
-                    panelType: 'window',
-                    type: 'application'
-                  }
-                })
-                actions = actions.concat(orderTabs)
-                tab.isAssociatedTabSequence = true
-                tab.tabsOrder = orderTabs
-              }
-            }
-
-            // get processess associated in tab
-            if (tabItem.processesList && tabItem.processesList.length) {
-              const processList = tabItem.processesList.map(processItem => {
-                // TODO: No list of parameters
-                // add process associated in vuex store
-                // dispatch('addProcessAssociated', {
-                //   processToGenerate: processItem,
-                //   containerUuidAssociated: tabItem.uuid
-                // })
-                return {
-                  id: processItem.id,
-                  uuid: processItem.uuid,
-                  name: processItem.name,
-                  type: 'process',
-                  panelType: 'process',
-                  description: processItem.description,
-                  help: processItem.help,
-                  isReport: processItem.isReport,
-                  isDirectPrint: processItem.isDirectPrint,
-                  containerUuidAssociated: tabItem.uuid,
-                  parentUuidAssociated: windowUuid,
-                  panelTypeAssociated: 'window'
-                }
-              })
-              actions = actions.concat(processList)
-            }
-            //  Add process menu
-            dispatch('setContextMenu', {
-              containerUuid: tab.uuid,
-              actions
-            })
-
-            if (tab.isParentTab) {
-              tabsListParent.push(tab)
-              return tab
-            }
-            if (!tab.isSortTab) {
-              tabsListChildren.push(tab)
-            }
-            return tab
-          })
-
-          const newWindow = {
-            ...responseWindow,
-            tabsList: tabs,
-            currentTab: tabsListParent[0],
-            tabsListParent,
-            tabsListChildren,
-            // app attributes
-            currentTabUuid: tabsListParent[0].uuid,
-            firstTab: tabsListParent[0],
-            firstTabUuid,
-            windowIndex: state.windowIndex + 1,
-            // App properties
-            isShowedTabsChildren: Boolean(tabsListChildren.length),
-            isShowedRecordNavigation: undefined,
-            isShowedAdvancedQuery: false
-          }
-          commit('addWindow', newWindow)
-          return newWindow
-        })
-        .catch(error => {
-          router.push({ path: '/dashboard' })
-          dispatch('tagsView/delView', routeToDelete)
-          showMessage({
-            message: language.t('login.unexpectedError'),
-            type: 'error'
-          })
-          console.warn(`Dictionary Window (State Window) - Error ${error.code}: ${error.message}.`)
-        })
-    },
-    getTabAndFieldFromServer({ dispatch, getters }, {
+    tableActionPerformed({ dispatch, getters }, {
       parentUuid,
       containerUuid,
-      panelType = 'window',
-      panelMetadata,
-      isAdvancedQuery = false
+      tableName,
+      columnName,
+      field,
+      value
     }) {
-      return new Promise(resolve => {
-        getTabMetadata(containerUuid)
-          .then(tabResponse => {
-            const additionalAttributes = {
-              parentUuid,
-              containerUuid,
-              isShowedFromUser: true,
-              panelType,
-              tableName: tabResponse.tableName,
-              //
-              isReadOnlyFromForm: false,
-              isAdvancedQuery,
-              isEvaluateValueChanges: !isAdvancedQuery
-            }
+      // console.log({
+      //   parentUuid,
+      //   containerUuid,
+      //   tableName,
+      //   columnName,
+      //   field,
+      //   value
+      // })
+    },
+    windowActionPerformed({ dispatch, getters }, {
+      parentUuid,
+      containerUuid,
+      tableName,
+      columnName,
+      field,
+      value
+    }) {
+      // console.log({
+      //   parentUuid,
+      //   containerUuid,
+      //   tableName,
+      //   columnName,
+      //   field,
+      //   value
+      // })
+      //  get value from store
+      if (!value) {
+        value = getters.getValueOfField({
+          containerUuid,
+          columnName
+        })
+      }
+      // request callouts
+      dispatch('runCallout', {
+        parentUuid,
+        containerUuid,
+        tableName,
+        columnName,
+        callout: field.callout,
+        oldValue: field.oldValue,
+        valueType: field.valueType,
+        value
+      })
+      //  Context Info
+      dispatch('reloadContextInfo', {
+        parentUuid,
+        containerUuid,
+        tableName,
+        columnName,
+        field
+      })
+      // For change options
+      if (fieldIsDisplayed(field)) {
+        dispatch('runServerAction', {
+          parentUuid,
+          containerUuid,
+          tableName,
+          columnName,
+          field
+        })
+      }
+    },
+    runServerAction({ dispatch, getters, commit }, {
+      parentUuid,
+      containerUuid,
+      tableName,
+      columnName,
+      field,
+      value
+    }) {
+      const uuid = getters.getValueOfField({
+        containerUuid,
+        columnName: 'UUID'
+      })
+      if (isEmptyValue(uuid)) {
+        dispatch('createNewEntity', {
+          parentUuid,
+          containerUuid
+        })
+          .then(() => {
+            // change old value so that it is not send in the next update
+            // commit('changeFieldValue', {
+            //   field,
+            //   newValue,
+            //   valueTo,
+            //   displayColumn,
+            //   isChangedOldValue: true
+            // })
+          })
+          .catch(error => {
+            showMessage({
+              message: error.message,
+              type: 'error'
+            })
+            console.warn(`Create Entity Error ${error.code}: ${error.message}.`)
+          })
+      } else {
+        dispatch('updateCurrentEntity', {
+          containerUuid,
+          recordUuid: uuid
+        })
+          .then(response => {
+            // change old value so that it is not send in the next update
+            showMessage({
+              message: language.t('notifications.updateFields') + field.name,
+              type: 'success'
+            })
+            // commit('changeFieldValue', {
+            //   field,
+            //   value,
+            //   valueTo,
+            //   displayColumn,
+            //   isChangedOldValue: true
+            // })
+            // change value in table
+            // dispatch('notifyRowTableChange', {
+            //   containerUuid,
+            //   row: response,
+            //   isEdit: false,
+            //   isParent: true
+            // })
+          })
+      }
+    },
+    reloadContextInfo({ dispatch, getters }, {
+      parentUuid,
+      containerUuid,
+      tableName,
+      columnName,
+      field
+    }) {
+      //  get value from store
+      const value = getters.getValueOfField({
+        containerUuid,
+        columnName
+      })
+      // request context info field
+      if (!isEmptyValue(value) && !isEmptyValue(field.contextInfo) && !isEmptyValue(field.contextInfo.sqlStatement)) {
+        let isSQL = false
+        let sqlStatement = field.contextInfo.sqlStatement
+        if (sqlStatement.includes('@')) {
+          if (sqlStatement.includes('@SQL=')) {
+            isSQL = true
+          }
+          sqlStatement = parseContext({
+            parentUuid,
+            containerUuid,
+            columnName,
+            value: sqlStatement,
+            isSQL
+          }).value
+          if (isSQL && String(sqlStatement) === '[object Object]') {
+            sqlStatement = sqlStatement.query
+          }
+        }
+        const contextInfo = dispatch.dispatch('getContextInfoValueFromServer', {
+          parentUuid,
+          containerUuid,
+          contextInfoUuid: field.contextInfo.uuid,
+          columnName,
+          sqlStatement
+        })
+        if (!isEmptyValue(contextInfo) && !isEmptyValue(contextInfo.messageText)) {
+          field.contextInfo.isActive = true
+          field.contextInfo.messageText.msgText = contextInfo.messageText
+          field.contextInfo.messageText.msgTip = contextInfo.messageTip
+        }
+      }
+    },
+    undoPanelToNew({ dispatch, rootGetters }, { containerUuid }) {
+      const oldAttributes = rootGetters.getColumnNamesAndValues({
+        containerUuid,
+        propertyName: 'oldValue',
+        isObjectReturn: true,
+        isAddDisplayColumn: true
+      })
+      dispatch('notifyPanelChange', {
+        containerUuid,
+        newValues: oldAttributes
+      })
+    },
+    createNewEntity({ commit, dispatch, getters, rootGetters }, {
+      parentUuid,
+      containerUuid
+    }) {
+      return new Promise((resolve, reject) => {
+        // exists some call to create new record with container uuid
+        if (getters.getInCreate(containerUuid)) {
+          return reject({
+            error: 0,
+            message: `In this panel (${containerUuid}) is a create new record in progress`
+          })
+        }
 
-            let isWithUuidField = false // indicates it contains the uuid field
-            let fieldLinkColumnName
-            //  Convert and add to app attributes
-            const fieldsList = tabResponse.fieldsList.map((fieldItem, index) => {
-              fieldItem = generateField({
-                fieldToGenerate: fieldItem,
-                moreAttributes: {
-                  ...additionalAttributes,
-                  fieldListIndex: index
-                }
-              })
+        const { tableName, fieldList } = rootGetters.getPanel(containerUuid)
+        // delete key from attributes
+        const attributesList = rootGetters.getColumnNamesAndValues({
+          containerUuid,
+          propertyName: 'value',
+          isEvaluateValues: true,
+          isAddDisplayColumn: false
+        })
 
-              if (!isWithUuidField && fieldItem.columnName === 'UUID') {
-                isWithUuidField = true
+        commit('addInCreate', {
+          containerUuid,
+          tableName,
+          attributesList
+        })
+        createEntity({
+          tableName,
+          attributesList
+        })
+          .then(createEntityResponse => {
+            const newValues = createEntityResponse.values
+            attributesList.forEach(element => {
+              if (element.columnName.includes('DisplayColumn')) {
+                newValues[element.columnName] = element.value
               }
-
-              if (fieldItem.isParent) {
-                fieldLinkColumnName = fieldItem.columnName
-              }
-
-              return fieldItem
             })
 
-            let isTabsChildren = false
-            if (!isAdvancedQuery) {
-              const window = getters.getWindow(parentUuid)
-              isTabsChildren = Boolean(window.tabsListChildren.length)
-            }
+            showMessage({
+              message: language.t('data.createRecordSuccessful'),
+              type: 'success'
+            })
 
-            if (!isWithUuidField) {
-              const fieldUuid = getFieldTemplate({
-                ...additionalAttributes,
-                isShowedFromUser: false,
-                name: 'UUID',
-                columnName: 'UUID',
-                componentPath: 'FieldText'
-              })
-              fieldsList.push(fieldUuid)
-            }
-
-            if (isEmptyValue(panelMetadata)) {
-              panelMetadata = getters.getTab(parentUuid, containerUuid)
-            }
-            //  Panel for save on store
-            const panel = {
-              ...panelMetadata,
-              isAdvancedQuery,
-              fieldLinkColumnName,
-              fieldList: fieldsList,
-              panelType,
-              // app attributes
-              isLoadFieldsList: true,
-              isShowedTotals: false,
-              isTabsChildren // to delete records assiciated
-            }
-
-            dispatch('addPanel', panel)
-            resolve(panel)
-
-            dispatch('changeTabAttribute', {
+            // update fields with new values
+            dispatch('notifyPanelChange', {
               parentUuid,
               containerUuid,
-              tab: panelMetadata,
-              attributeName: 'isLoadFieldsList',
-              attributeValue: true
+              newValues,
+              isSendToServer: false
+            })
+            dispatch('addNewRow', {
+              parentUuid,
+              containerUuid,
+              isPanelValues: true,
+              isEdit: false
+            })
+
+            // set data log to undo action
+            const fieldId = fieldList.find(itemField => itemField.isKey)
+            dispatch('setDataLog', {
+              containerUuid,
+              tableName,
+              recordId: fieldId.value, // TODO: Verify performance with tableName_ID
+              recordUuid: newValues.UUID,
+              eventType: 'INSERT'
+            })
+
+            const oldRoute = router.app._route
+            router.push({
+              name: oldRoute.name,
+              params: {
+                ...oldRoute.params
+              },
+              query: {
+                ...oldRoute.query,
+                action: createEntityResponse.uuid
+              }
+            })
+            dispatch('tagsView/delView', oldRoute, true)
+
+            resolve({
+              data: newValues,
+              recordUuid: createEntityResponse.uuid,
+              recordId: createEntityResponse.id,
+              tableName: createEntityResponse.tableName
             })
           })
           .catch(error => {
             showMessage({
-              message: language.t('login.unexpectedError'),
+              message: error.message,
               type: 'error'
             })
-            console.warn(`Dictionary Tab (State Window) - Error ${error.code}: ${error.message}.`)
+            console.warn(`Create Entity error: ${error.message}.`)
+            reject(error)
+          })
+          .finally(() => {
+            commit('deleteInCreate', {
+              containerUuid,
+              tableName,
+              attributesList
+            })
           })
       })
     },
-    setCurrentTab({ commit, getters }, {
+    createEntityFromTable({ commit, dispatch, getters, rootGetters }, {
       parentUuid,
       containerUuid,
-      window,
+      row
+    }) {
+      // exists some call to create new record with container uuid
+      if (getters.getInCreate(containerUuid)) {
+        return {
+          error: 0,
+          message: `In this panel (${containerUuid}) is a create new record in progress.`
+        }
+      }
+      const { tableName, isParentTab } = rootGetters.getPanel(containerUuid)
+
+      // TODO: Add support to Binary columns (BinaryData)
+      const columnsToDontSend = ['BinaryData', 'isEdit', 'isNew', 'isSendServer']
+
+      // TODO: Evaluate peformance without filter using delete(prop) before convert object to array
+      // attributes or fields
+      const fieldsList = getters.getFieldsListFromPanel(containerUuid)
+      const attributesList = []
+      fieldsList.forEach(itemAttribute => {
+        if (columnsToDontSend.includes(itemAttribute.columnName) || itemAttribute.columnName.includes('DisplayColumn')) {
+          return false
+        }
+        if (isEmptyValue(row[itemAttribute.columnName])) {
+          return false
+        }
+
+        attributesList.push({
+          value: row[itemAttribute.columnName],
+          columnName: itemAttribute.columnName,
+          valueType: itemAttribute.valueType
+        })
+      })
+
+      commit('addInCreate', {
+        containerUuid,
+        tableName,
+        attributesList
+      })
+
+      let isError = false
+      return createEntity({
+        tableName,
+        attributesList
+      })
+        .then(createEntityResponse => {
+          showMessage({
+            message: language.t('data.createRecordSuccessful'),
+            type: 'success'
+          })
+          if (isParentTab) {
+            // redirect to create new record
+            const oldRoute = router.app._route
+            router.push({
+              name: oldRoute.name,
+              params: {
+                ...oldRoute.params
+              },
+              query: {
+                ...oldRoute.query,
+                action: createEntityResponse.recordUuid
+              }
+            })
+          }
+          return {
+            data: createEntityResponse.values,
+            ...createEntityResponse
+          }
+        })
+        .catch(error => {
+          showMessage({
+            message: error.message,
+            type: 'error'
+          })
+          console.warn(`Create Entity Table Error ${error.code}: ${error.message}.`)
+          isError = true
+        })
+        .finally(() => {
+          if (isError) {
+            dispatch('addNewRow', {
+              containerUuid,
+              row
+            })
+          } else {
+            // refresh record list
+            dispatch('getDataListTab', {
+              parentUuid,
+              containerUuid
+            })
+              .catch(error => {
+                console.warn(`Error getting data list tab. Message: ${error.message}, code ${error.code}.`)
+              })
+          }
+          commit('deleteInCreate', {
+            containerUuid,
+            tableName,
+            attributesList
+          })
+        })
+    },
+    updateCurrentEntity({ dispatch, rootGetters }, {
+      containerUuid,
+      recordUuid = null
+    }) {
+      const panel = rootGetters.getPanel(containerUuid)
+      if (!recordUuid) {
+        recordUuid = rootGetters.getUuid(containerUuid)
+      }
+
+      // TODO: Add support to Binary columns (BinaryData)
+      const columnsToDontSend = ['Account_Acct']
+
+      // attributes or fields
+      let finalAttributes = rootGetters.getColumnNamesAndValues({
+        containerUuid: containerUuid,
+        isEvaluatedChangedValue: true
+      })
+
+      finalAttributes = finalAttributes.filter(itemAttribute => {
+        if (columnsToDontSend.includes(itemAttribute.columnName) || itemAttribute.columnName.includes('DisplayColumn')) {
+          return false
+        }
+        const field = panel.fieldList.find(itemField => itemField.columnName === itemAttribute.columnName)
+        if (!field || !field.isUpdateable || !field.isDisplayed) {
+          return false
+        }
+        return true
+      })
+      return updateEntity({
+        tableName: panel.tableName,
+        recordUuid,
+        attributesList: finalAttributes
+      })
+        .then(updateEntityResponse => {
+          const newValues = updateEntityResponse.values
+          // set data log to undo action
+          // TODO: Verify performance with tableName_ID
+          let recordId = updateEntityResponse.id
+          if (isEmptyValue(recordId)) {
+            recordId = newValues[`${panel.tableName}_ID`]
+          }
+          if (isEmptyValue(recordId)) {
+            const fieldId = panel.fieldList.find(itemField => itemField.isKey)
+            recordId = fieldId.value
+          }
+
+          if (isEmptyValue(recordUuid)) {
+            recordUuid = updateEntityResponse.uuid
+          }
+          if (isEmptyValue(recordUuid)) {
+            recordUuid = newValues.UUID
+          }
+
+          dispatch('setDataLog', {
+            containerUuid,
+            tableName: panel.tableName,
+            recordId,
+            recordUuid,
+            eventType: 'UPDATE'
+          })
+          if (rootGetters.getShowContainerInfo) {
+            dispatch('listRecordLogs', {
+              tableName: panel.tableName,
+              recordId
+            })
+          }
+          return newValues
+        })
+        .catch(error => {
+          showMessage({
+            message: error.message,
+            type: 'error'
+          })
+          console.warn(`Update Entity Error ${error.code}: ${error.message}`)
+        })
+    },
+    updateCurrentEntityFromTable({ rootGetters }, {
+      containerUuid,
+      row
+    }) {
+      const { tableName, fieldList } = rootGetters.getPanel(containerUuid)
+
+      // TODO: Add support to Binary columns (BinaryData)
+      const columnsToDontSend = ['BinaryData', 'isEdit', 'isNew', 'isSendServer']
+
+      // TODO: Evaluate peformance without filter using delete(prop) before convert object to array
+      // attributes or fields
+      let finalAttributes = convertObjectToArrayPairs(row)
+      finalAttributes = finalAttributes.filter(itemAttribute => {
+        if (columnsToDontSend.includes(itemAttribute.columnName) || itemAttribute.columnName.includes('DisplayColumn')) {
+          return false
+        }
+        const field = fieldList.find(itemField => itemField.columnName === itemAttribute.columnName)
+        if (!field || !field.isUpdateable || !field.isDisplayed) {
+          return false
+        }
+        return true
+      })
+
+      return updateEntity({
+        tableName,
+        recordUuid: row.UUID,
+        attributesList: finalAttributes
+      })
+        .then(response => {
+          return response
+        })
+        .catch(error => {
+          showMessage({
+            message: error.message,
+            type: 'error'
+          })
+          console.warn(`Update Entity Table Error ${error.code}: ${error.message}.`)
+        })
+    },
+    /**
+     * Update record after run process associated with window
+     * @param {string} parentUuid
+     * @param {string} containerUuid
+     * @param {object} tab
+     */
+    updateRecordAfterRunProcess({ dispatch, rootGetters }, {
+      parentUuid,
+      containerUuid,
       tab
     }) {
-      if (isEmptyValue(window)) {
-        window = getters.getWindow(parentUuid)
-      }
-      if (isEmptyValue(tab)) {
-        tab = window.tabsList.find(itemTab => itemTab.uuid === containerUuid)
-      }
-
-      commit('setCurrentTab', {
-        window,
-        tab
-      })
-    },
-    changeWindowAttribute({ commit, getters }, {
-      parentUuid,
-      window,
-      attributeName,
-      attributeNameControl,
-      attributeValue
-    }) {
-      if (isEmptyValue(window)) {
-        window = getters.getWindow(parentUuid)
-      }
-
-      commit('changeWindowAttribute', {
+      const recordUuid = rootGetters.getUuid(containerUuid)
+      // get new values
+      dispatch('getEntity', {
         parentUuid,
-        window,
-        attributeName,
-        attributeNameControl,
-        attributeValue
+        containerUuid,
+        tableName: tab.tableName,
+        recordUuid
       })
+        .then(response => {
+          // update panel
+          if (tab.isParentTab) {
+            dispatch('notifyPanelChange', {
+              parentUuid,
+              containerUuid,
+              newValues: response,
+              isSendCallout: false,
+              isSendToServer: false
+            })
+          }
+          // update row in table
+          dispatch('notifyRowTableChange', {
+            parentUuid,
+            containerUuid,
+            row: response,
+            isEdit: false
+          })
+        })
     },
-    changeTabAttribute({ commit, getters }, {
+    deleteEntity({ dispatch, rootGetters }, {
       parentUuid,
       containerUuid,
-      tab,
-      attributeName,
-      attributeNameControl,
-      attributeValue
+      recordUuid,
+      recordId,
+      row
     }) {
-      if (isEmptyValue(tab)) {
-        tab = getters.getTab(parentUuid, containerUuid)
+      return new Promise(resolve => {
+        const panel = rootGetters.getPanel(containerUuid)
+        if (!isEmptyValue(row)) {
+          recordUuid = row.UUID
+          recordId = row[`${panel.tableName}_ID`]
+        }
+
+        deleteEntity({
+          tableName: panel.tableName,
+          recordUuid,
+          recordId
+        })
+          .then(responseDeleteEntity => {
+            // refresh record list
+            dispatch('getDataListTab', {
+              parentUuid,
+              containerUuid
+            })
+              .then(responseDataList => {
+                if (panel.isParentTab) {
+                  // if response is void, go to new record
+                  if (responseDataList.length <= 0) {
+                    dispatch('resetPanelToNew', {
+                      parentUuid,
+                      containerUuid,
+                      panelType: 'window',
+                      isNewRecord: true
+                    })
+                  } else {
+                    const oldRoute = router.app._route
+                    // else display first record of table in panel
+                    router.push({
+                      name: oldRoute.name,
+                      params: {
+                        ...oldRoute.params
+                      },
+                      query: {
+                        ...oldRoute.query,
+                        action: responseDataList[0].UUID
+                      }
+                    })
+                  }
+                }
+              })
+              .catch(error => {
+                console.warn(`Error getting data list tab. Message: ${error.message}, code ${error.code}.`)
+              })
+            showMessage({
+              message: language.t('data.deleteRecordSuccessful'),
+              type: 'success'
+            })
+
+            if (isEmptyValue(recordId)) {
+              // TODO: Verify performance with tableName_ID
+              const fieldId = panel.fieldList.find(itemField => itemField.isKey)
+              recordId = fieldId.value
+            }
+            // set data log to undo action
+            dispatch('setDataLog', {
+              containerUuid,
+              tableName: panel.tableName,
+              recordId,
+              recordUuid,
+              eventType: 'DELETE'
+            })
+
+            resolve(responseDeleteEntity)
+          })
+          .catch(error => {
+            showMessage({
+              message: language.t('data.deleteRecordError'),
+              type: 'error'
+            })
+            console.warn(`Delete Entity - Error ${error.message}, Code: ${error.code}.`)
+          })
+      })
+    },
+    /**
+     * Delete selection records in table
+     * @param {string} parentUuid
+     * @param {string} containerUuid
+     * @param {string} tableName
+     * @param {boolean} isParentTab
+     */
+    deleteSelectionDataList({ dispatch, rootGetters }, {
+      parentUuid,
+      containerUuid,
+      tableName,
+      isParentTab
+    }) {
+      if (isEmptyValue(tableName) || isEmptyValue(isParentTab)) {
+        const tab = rootGetters.getTab(parentUuid, containerUuid)
+        tableName = tab.tableName
+        isParentTab = tab.isParentTab
       }
-      commit('changeTabAttribute', {
-        tab,
-        attributeName,
-        attributeValue,
-        attributeNameControl
+      const allData = rootGetters.getDataRecordAndSelection(containerUuid)
+      let selectionLength = allData.selection.length
+
+      allData.selection.forEach((record, index) => {
+        // validate if the registry row has no uuid before sending to the server
+        if (isEmptyValue(record.UUID)) {
+          selectionLength = selectionLength - 1
+          console.warn(`This row does not contain a record with UUID`, record)
+          // refresh record list
+          dispatch('getDataListTab', {
+            parentUuid,
+            containerUuid
+          })
+            .catch(error => {
+              console.warn(`Error getting data list tab. Message: ${error.message}, code ${error.code}.`)
+            })
+          return
+        }
+        deleteEntity({
+          tableName,
+          recordUuid: record.UUID
+        })
+          .then(() => {
+            if (isParentTab) {
+              // redirect to create new record
+              const oldRoute = router.app._route
+              if (record.UUID === oldRoute.query.action) {
+                router.push({
+                  name: oldRoute.name,
+                  params: {
+                    ...oldRoute.params
+                  },
+                  query: {
+                    ...oldRoute.query,
+                    action: 'create-new'
+                  }
+                })
+                // clear fields with default values
+                dispatch('resetPanelToNew', {
+                  parentUuid,
+                  containerUuid
+                })
+                // delete view with uuid record delete
+                dispatch('tagsView/delView', oldRoute, true)
+              }
+            }
+
+            if ((index + 1) >= selectionLength) {
+              // refresh record list
+              dispatch('getDataListTab', {
+                parentUuid,
+                containerUuid
+              })
+                .catch(error => {
+                  console.warn(`Error getting data list tab. Message: ${error.message}, code ${error.code}.`)
+                })
+              showMessage({
+                message: language.t('data.deleteRecordSuccessful'),
+                type: 'success'
+              })
+            }
+          })
+      })
+    },
+    undoModifyData({ getters }, {
+      containerUuid,
+      recordUuid
+    }) {
+      return rollbackEntity(getters.getDataLog(containerUuid, recordUuid))
+        .then(response => {
+          return response
+        })
+        .catch(error => {
+          showMessage({
+            message: error.message,
+            type: 'error'
+          })
+          console.warn(`Rollback Entity error: ${error.message}. Code: ${error.code}.`)
+        })
+    },
+    setDataLog({ commit }, {
+      containerUuid,
+      tableName,
+      recordId,
+      recordUuid,
+      eventType
+    }) {
+      commit('setDataLog', {
+        containerUuid,
+        tableName,
+        recordId,
+        recordUuid,
+        eventType
+      })
+    },
+    /**
+     * Get data to table in tab
+     * @param {string}  parentUuid, window to search record data
+     * @param {string}  containerUuid, tab to search record data
+     * @param {string}  recordUuid, uuid to search
+     * @param {boolean} isRefreshPanel, if main panel is updated with new response data
+     * @param {boolean} isLoadAllRecords, if main panel is updated with new response data
+     */
+    getDataListTab({ dispatch, rootGetters }, {
+      parentUuid,
+      containerUuid,
+      recordUuid,
+      referenceWhereClause = '',
+      columnName,
+      value,
+      criteria,
+      isAddRecord = false,
+      isLoadAllRecords = false,
+      isRefreshPanel = false,
+      isReference = false,
+      isShowNotification = true
+    }) {
+      const tab = rootGetters.getTab(parentUuid, containerUuid)
+
+      let parsedQuery = tab.query
+      if (!isEmptyValue(parsedQuery) && parsedQuery.includes('@')) {
+        parsedQuery = parseContext({
+          parentUuid,
+          containerUuid,
+          value: tab.query,
+          isBooleanToString: true
+        }).value
+      }
+
+      let parsedWhereClause = tab.whereClause
+      if (!isEmptyValue(parsedWhereClause) && parsedWhereClause.includes('@')) {
+        parsedWhereClause = parseContext({
+          parentUuid,
+          containerUuid,
+          value: tab.whereClause,
+          isBooleanToString: true
+        }).value
+      }
+
+      if (isReference) {
+        if (!isEmptyValue(parsedWhereClause)) {
+          parsedWhereClause += ` AND ${referenceWhereClause}`
+        } else {
+          parsedWhereClause += referenceWhereClause
+        }
+      }
+
+      if (!isEmptyValue(criteria)) {
+        if (!isEmptyValue(parsedWhereClause)) {
+          parsedWhereClause += ` AND ${criteria.whereClause}`
+        } else {
+          parsedWhereClause += criteria.whereClause
+        }
+      }
+
+      const conditionsList = []
+      // TODO: evaluate if overwrite values to conditions
+      if (!isLoadAllRecords && tab.isParentTab && !isEmptyValue(tab.tableName) && !isEmptyValue(value)) {
+        conditionsList.push({
+          columnName,
+          value
+        })
+      }
+      return dispatch('getObjectListFromCriteria', {
+        parentUuid,
+        containerUuid,
+        tableName: tab.tableName,
+        query: parsedQuery,
+        whereClause: parsedWhereClause,
+        orderByClause: tab.orderByClause,
+        conditionsList,
+        isParentTab: tab.isParentTab,
+        isAddRecord,
+        isShowNotification
+      })
+        .then(response => {
+          if (isRefreshPanel && !isEmptyValue(recordUuid) && recordUuid !== 'create-new') {
+            const newValues = response.find(itemData => itemData.UUID === recordUuid)
+            if (newValues) {
+              // update fields with values obtained from the server
+              dispatch('notifyPanelChange', {
+                parentUuid,
+                containerUuid,
+                newValues,
+                isSendCallout: false,
+                isSendToServer: false
+              })
+            } else {
+              // this record is missing (Deleted or the query does not include it)
+              dispatch('resetPanelToNew', {
+                parentUuid,
+                containerUuid
+              })
+            }
+          }
+          return response
+        })
+        .catch(error => {
+          return error
+        })
+        .finally(() => {
+          const currentData = rootGetters.getDataRecordAndSelection(containerUuid)
+          const { originalNextPageToken, pageNumber, recordCount } = currentData
+          let nextPage = pageNumber
+          const isAdd = isAddRecord
+          if (originalNextPageToken && isAddRecord) {
+            const pageInToken = originalNextPageToken.substring(originalNextPageToken.length - 2)
+            if (pageInToken === '-1') {
+              isAddRecord = false
+            }
+            if (pageNumber === 1 && recordCount > 50) {
+              nextPage = nextPage + 1
+              isAddRecord = true
+            }
+          } else {
+            isAddRecord = false
+          }
+          if (recordCount <= 50) {
+            isAddRecord = false
+          }
+
+          if (isAddRecord) {
+            dispatch('setPageNumber', {
+              parentUuid,
+              containerUuid,
+              pageNumber: nextPage,
+              panelType: 'window',
+              isAddRecord,
+              isShowNotification: false
+            })
+          }
+          if (isAdd && isAdd !== isAddRecord) {
+            if (tab.isSortTab) {
+              const record = rootGetters.getDataRecordsList(containerUuid)
+              const recordToTab = record
+                .map(itemRecord => {
+                  return {
+                    ...itemRecord
+                  }
+                })
+                .sort((itemA, itemB) => {
+                  return itemA[tab.sortOrderColumnName] - itemB[tab.sortOrderColumnName]
+                })
+              dispatch('setTabSequenceRecord', recordToTab)
+            }
+          }
+        })
+    },
+    /**
+     * Get references asociate to record
+     * @param {string} parentUuid as windowUuid
+     * @param {string} containerUuid
+     * @param {string} tableName
+     * @param {string} recordUuid
+     */
+    getReferencesListFromServer({ commit, rootGetters }, {
+      parentUuid: windowUuid,
+      containerUuid,
+      tableName,
+      recordUuid
+    }) {
+      if (isEmptyValue(tableName)) {
+        tableName = rootGetters.getTab(windowUuid, containerUuid).tableName
+      }
+      return new Promise(resolve => {
+        getReferencesList({
+          windowUuid,
+          tableName,
+          recordUuid
+        })
+          .then(referenceResponse => {
+            const referencesList = referenceResponse.referencesList.map(item => {
+              return {
+                ...item,
+                recordUuid,
+                type: 'reference'
+              }
+            })
+            const references = {
+              ...referenceResponse,
+              windowUuid,
+              recordUuid,
+              referencesList
+            }
+            commit('addReferencesList', references)
+            resolve(referenceResponse)
+          })
+          .catch(error => {
+            console.warn(`References Load Error ${error.code}: ${error.message}.`)
+          })
+      })
+    },
+    setWindowOldRoute({ commit }, oldPath = { path: '', fullPath: '', query: {}}) {
+      commit('setWindowOldRoute', oldPath)
+    },
+    setTabSequenceRecord({ commit }, record) {
+      commit('setTabSequenceRecord', record)
+    },
+    /**
+     * Update records in tab sort
+     * @param {string} containerUuid
+     * @param {string} parentUuid
+     */
+    updateSequence({ state, commit, dispatch, getters, rootGetters }, {
+      parentUuid,
+      containerUuid
+    }) {
+      const { tableName, sortOrderColumnName, sortYesNoColumnName, tabAssociatedUuid } = rootGetters.getTab(parentUuid, containerUuid)
+      const listSequenceToSet = getters.getTabSequenceRecord
+      const recordData = rootGetters.getDataRecordsList(containerUuid)
+
+      // scrolls through the logs and checks if there is a change to be sent to server
+      recordData.forEach(itemData => {
+        const dataSequence = listSequenceToSet.find(item => item.UUID === itemData.UUID)
+        if (itemData[sortOrderColumnName] === dataSequence[sortOrderColumnName]) {
+          return
+        }
+        const valuesToSend = [{
+          columnName: sortOrderColumnName,
+          value: dataSequence[sortOrderColumnName]
+        }]
+
+        if (itemData[sortYesNoColumnName] !== dataSequence[sortYesNoColumnName]) {
+          valuesToSend.push({
+            columnName: sortYesNoColumnName,
+            value: dataSequence[sortYesNoColumnName]
+          })
+        }
+
+        const countRequest = state.totalRequest + 1
+        commit('setTotalRequest', countRequest)
+
+        updateEntity({
+          tableName,
+          recordUuid: itemData.UUID,
+          attributesList: valuesToSend
+        })
+          .catch(error => {
+            showMessage({
+              message: error.message,
+              type: 'error'
+            })
+            console.warn(`Update Entity Table Error ${error.code}: ${error.message}`)
+          })
+          .finally(() => {
+            const countResponse = state.totalResponse + 1
+            commit('setTotalResponse', countResponse)
+            if (state.totalResponse === state.totalRequest) {
+              showMessage({
+                message: language.t('notifications.updateSuccessfully'),
+                type: 'success'
+              })
+              dispatch('setShowDialog', {
+                type: 'window',
+                action: undefined
+              })
+              commit('setTotalRequest', 0)
+              commit('setTotalResponse', 0)
+
+              dispatch('setRecordSelection', {
+                parentUuid,
+                containerUuid,
+                isLoaded: false
+              })
+              dispatch('setTabSequenceRecord', [])
+
+              // refresh record list in table source
+              dispatch('getDataListTab', {
+                parentUuid,
+                containerUuid: tabAssociatedUuid
+              })
+                .catch(error => {
+                  console.warn(`Error getting data list tab. Message: ${error.message}, code ${error.code}.`)
+                })
+            }
+          })
       })
     }
   },
   getters: {
-    getWindow: (state) => (windowUuid) => {
-      return state.window.find(
-        item => item.uuid === windowUuid
-      )
+    getInCreate: (state) => (containerUuid) => {
+      return state.inCreate.find(item => item.containerUuid === containerUuid)
     },
-    getIsShowedRecordNavigation: (state, getters) => (windowUuid) => {
-      const window = getters.getWindow(windowUuid)
-      if (window) {
-        return window.isShowedRecordNavigation
-      }
-      return window
+    getReferencesList: (state) => (windowUuid, recordUuid) => {
+      return state.references.find(item => item.windowUuid === windowUuid && item.recordUuid === recordUuid)
     },
-    getTab: (state, getters) => (windowUuid, tabUuid) => {
-      const window = getters.getWindow(windowUuid)
-      if (window) {
-        return window.tabsList.find(tabItem => {
-          return tabItem.uuid === tabUuid
-        })
-      }
-      return window
+    getReferencesInfo: (state, getters) => ({ windowUuid, recordUuid, referenceUuid }) => {
+      const references = getters.getReferencesList(windowUuid, recordUuid)
+      return references.referencesList.find(item => item.uuid === referenceUuid)
     },
-    getCurrentTab: (state, getters) => (windowUuid) => {
-      const window = getters.getWindow(windowUuid)
-      if (window) {
-        return window.tabsList.find(tabItem => {
-          return tabItem.uuid === window.currentTabUuid
-        })
-      }
-      return {
-        isInsertRecord: false
-      }
+    getTabSequenceRecord: (state) => {
+      return state.tabSequenceRecord
     },
-    getTableNameFromTab: (state, getters) => (windowUuid, tabUuid) => {
-      return getters.getTab(windowUuid, tabUuid).tableName
+    getDataLog: (state) => (containerUuid, recordUuid) => {
+      const current = state.dataLog
+      if (current.containerUuid === containerUuid &&
+        ((current.recordUuid === recordUuid) ||
+        (current.eventType === 'DELETE' && recordUuid === 'create-new'))) {
+        return current
+      }
+      return undefined
     }
   }
 }
 
-export default window
+export default windowControl
